refactor(mpwd): Modularize application structure
Extracted file scanning, copying, cleaning, and progress management logic into dedicated modules.
Moved common types and error definitions to a new `types` module.
Reduced `main.rs` to primarily handle argument parsing and command dispatch.
