+++ rust/move-photos-without-duplicates/src/main.rs
@@ -1,41 +1,19 @@
+mod cleaner;
+mod copier;
 mod db;
+mod progress;
+mod scanner;
+mod types;
 
 use camino::Utf8PathBuf;
 use clap::{Parser, Subcommand};
-use db::{Database, CopiedFileRecord, DuplicateFileRecord};
+use cleaner::FileCleaner;
+use copier::FileCopier;
+use db::Database;
 use error_stack::{Result, ResultExt};
-use indicatif::{MultiProgress, ProgressBar, ProgressStyle};
-use rayon::prelude::*;
-use sha2::{Digest, Sha256};
-use std::fs;
-use std::sync::Arc;
-use std::time::SystemTime;
-use thiserror::Error;
-use walkdir::WalkDir;
-
-#[derive(Debug, Error)]
-pub enum AppError {
-    #[error("Database error")]
-    DB,
-
-    #[error("IO error")]
-    IO,
-
-    #[error("File processing error")]
-    FileProcessing,
-
-    #[error("Directory traversal error")]
-    DirectoryTraversal,
-
-    #[error("Validation error")]
-    Validation,
-
-    #[error("Copy operation error")]
-    Copy,
-
-    #[error("Cleanup operation error")]
-    Cleanup,
-}
+use progress::ProgressManager;
+use scanner::FileScanner;
+use types::AppError;
 
 #[derive(Parser)]
 #[command(name = "move-photos-without-duplicates")]
@@ -85,789 +63,12 @@ enum Commands {
     },
 }
 
-#[derive(Debug, Clone)]
-struct FileInfo {
-    path: Utf8PathBuf,
-    hash: String,
-    size: u64,
-    last_modified: SystemTime,
-}
-
-fn calculate_file_hash(file_path: &Utf8PathBuf) -> Result<String, AppError> {
-    let contents = fs::read(file_path)
-        .change_context(AppError::IO)
-        .attach_printable_lazy(|| format!("Failed to read file contents: {}", file_path))?;
-
-    let mut hasher = Sha256::new();
-    hasher.update(&contents);
-    let hash = hasher.finalize();
-
-    Ok(format!("{:x}", hash))
-}
-
-fn process_file(entry_path: &Utf8PathBuf) -> Result<FileInfo, AppError> {
-    let metadata = fs::metadata(entry_path)
-        .change_context(AppError::IO)
-        .attach_printable_lazy(|| {
-            format!(
-                "Failed to get metadata (stat operation) for: {}",
-                entry_path
-            )
-        })?;
-
-    let hash = calculate_file_hash(entry_path)?;
-    let size = metadata.len();
-
-    let last_modified = metadata
-        .modified()
-        .change_context(AppError::IO)
-        .attach_printable_lazy(|| {
-            format!(
-                "Failed to get modification time (stat.st_mtime) for: {}",
-                entry_path
-            )
-        })?;
-
-    Ok(FileInfo {
-        path: entry_path.clone(),
-        hash,
-        size,
-        last_modified,
-    })
-}
-
-fn setup_progress_logging() -> MultiProgress {
-    let multi = MultiProgress::new();
-    
-    // For now, just return MultiProgress - we'll keep logging simple
-    multi
-}
-
-async fn scan_and_hash_directory(directory: &Utf8PathBuf, db: &Database, multi: &MultiProgress, batch_size: usize) -> Result<(), AppError> {
-    // Phase 1: Directory scanning with spinner
-    let scan_pb = multi.add(ProgressBar::new_spinner());
-    scan_pb.set_style(
-        ProgressStyle::with_template("🔍 {spinner:.green} {wide_msg}")
-            .unwrap()
-            .tick_chars("⠁⠂⠄⡀⢀⠠⠐⠈ "),
-    );
-    scan_pb.set_message(format!("Scanning directory: {}", directory));
-
-    // Collect all entries and handle walkdir errors
-    let mut file_paths = Vec::new();
-    let mut traversal_errors = 0;
-
-    for entry in WalkDir::new(directory) {
-        scan_pb.tick();
-        match entry {
-            Ok(dir_entry) => {
-                if dir_entry.file_type().is_file() {
-                    match Utf8PathBuf::from_path_buf(dir_entry.path().to_path_buf()) {
-                        Ok(utf8_path) => {
-                            file_paths.push(utf8_path);
-                            scan_pb.set_message(format!("Scanning... found {} files", file_paths.len()));
-                        },
-                        Err(path_buf) => {
-                            eprintln!(
-                                "ERROR: Non-UTF8 path encountered during path conversion: {:?}",
-                                path_buf
-                            );
-                            traversal_errors += 1;
-                        }
-                    }
-                }
-            }
-            Err(err) => {
-                eprintln!("ERROR: Directory traversal failed");
-                if let Some(path) = err.path() {
-                    eprintln!("  Path: {:?}", path);
-                }
-                if let Some(io_err) = err.io_error() {
-                    eprintln!(
-                        "  IO Error during directory read/opendir operation: {}",
-                        io_err
-                    );
-                } else {
-                    eprintln!("  Error: {}", err);
-                }
-                traversal_errors += 1;
-            }
-        }
-    }
-
-    if traversal_errors > 0 {
-        eprintln!(
-            "WARNING: Encountered {} errors during directory traversal",
-            traversal_errors
-        );
-    }
-
-    scan_pb.finish_with_message(format!("✓ Found {} files to process", file_paths.len()));
-
-    // Phase 2: Batched file processing with progress bar
-    let process_pb = multi.add(ProgressBar::new(file_paths.len() as u64));
-    process_pb.set_style(
-        ProgressStyle::with_template("📁 [{elapsed}] {bar:40.cyan/blue} {pos:>7}/{len:7} {msg}")
-            .unwrap()
-            .progress_chars("##-"),
-    );
-    process_pb.set_message("Processing files in batches...");
-
-    let process_pb_arc = Arc::new(process_pb);
-    let total_files = file_paths.len();
-    let num_batches = (total_files + batch_size - 1) / batch_size;
-    
-    let mut total_processing_errors = 0;
-    let mut total_db_errors = 0;
-    let mut total_processed = 0;
-
-    println!("Processing {} files in {} batches of {} files each", total_files, num_batches, batch_size);
-
-    // Process files in batches
-    for (batch_num, batch_paths) in file_paths.chunks(batch_size).enumerate() {
-        let batch_start = batch_num * batch_size;
-        process_pb_arc.set_message(format!("Processing batch {}/{} (files {}-{})", 
-            batch_num + 1, num_batches, batch_start + 1, batch_start + batch_paths.len()));
-
-        // Process files in this batch in parallel
-        let batch_results: Vec<Result<FileInfo, AppError>> = batch_paths
-            .par_iter()
-            .map(|path| {
-                let result = process_file(path);
-                process_pb_arc.inc(1);
-                result
-            })
-            .collect();
-
-        // Separate successful results from errors
-        let mut batch_file_infos = Vec::new();
-        let mut batch_processing_errors = 0;
-
-        for (path, result) in batch_paths.iter().zip(batch_results.iter()) {
-            match result {
-                Ok(info) => {
-                    batch_file_infos.push(info.clone());
-                }
-                Err(e) => {
-                    eprintln!("ERROR: Failed to process file: {}", path);
-                    eprintln!("  Details: {:?}", e);
-                    batch_processing_errors += 1;
-                }
-            }
-        }
-
-        total_processing_errors += batch_processing_errors;
-        total_processed += batch_file_infos.len();
-
-        // Store this batch in database using transaction
-        if !batch_file_infos.is_empty() {
-            match db.batch_upsert_file_hashes(&batch_file_infos).await {
-                Ok(()) => {
-                    process_pb_arc.set_message(format!("✓ Batch {}/{} stored ({} files)", 
-                        batch_num + 1, num_batches, batch_file_infos.len()));
-                }
-                Err(e) => {
-                    eprintln!("ERROR: Failed to store batch {} in database", batch_num + 1);
-                    eprintln!("  Details: {:?}", e);
-                    total_db_errors += batch_file_infos.len();
-                }
-            }
-        }
-
-        // Brief pause between batches to prevent overwhelming the system
-        if batch_num < num_batches - 1 {
-            tokio::time::sleep(tokio::time::Duration::from_millis(10)).await;
-        }
-    }
-
-    process_pb_arc.finish_with_message(format!("✓ Processed {} files in {} batches", total_processed, num_batches));
-
-    // Report summary
-    let total_errors = traversal_errors + total_processing_errors + total_db_errors;
-    if total_errors > 0 {
-        eprintln!("SUMMARY: Completed with {} total errors", total_errors);
-        eprintln!("  - Directory traversal errors: {}", traversal_errors);
-        eprintln!("  - File processing errors: {}", total_processing_errors);
-        eprintln!("  - Database storage errors: {}", total_db_errors);
-    } else {
-        println!("SUMMARY: All {} files processed successfully in {} batches!", total_processed, num_batches);
-    }
-
-    Ok(())
-}
-
-fn validate_copy_command(
-    source: &Utf8PathBuf,
-    target: &Utf8PathBuf,
-) -> Result<(), AppError> {
-    // Check that both paths are directories
-    if !source.is_dir() {
-        return Err(
-            error_stack::Report::new(AppError::Validation).attach_printable(format!(
-                "Source directory does not exist or is not a directory: {}",
-                source
-            )),
-        );
-    }
-
-    if !target.is_dir() {
-        return Err(
-            error_stack::Report::new(AppError::Validation).attach_printable(format!(
-                "Target directory does not exist or is not a directory: {}",
-                target
-            )),
-        );
-    }
-
-    println!("Validation passed:");
-    println!("  Source directory: {}", source);
-    println!("  Target directory: {}", target);
-    
-    // Print warning about target directory needing to be hashed
-    println!("\n⚠️  WARNING: The target directory should be hashed before running copy operations");
-    println!("   to ensure effective duplicate detection. Run:");
-    println!("   {} hash {}", env!("CARGO_PKG_NAME"), target);
-
-    Ok(())
-}
-
-async fn copy_files_without_duplicates(
-    dir_a: &Utf8PathBuf,
-    dir_b: &Utf8PathBuf,
-    db: &Database,
-    multi: &MultiProgress,
-    dry_run: bool,
-) -> Result<(), AppError> {
-    println!("Starting copy operation from {} to {}", dir_a, dir_b);
-    if dry_run {
-        println!("DRY RUN MODE: No files will actually be copied");
-    }
-
-    // Phase 1: Directory scanning with spinner
-    let scan_pb = multi.add(ProgressBar::new_spinner());
-    scan_pb.set_style(
-        ProgressStyle::with_template("🔍 {spinner:.green} {wide_msg}")
-            .unwrap()
-            .tick_chars("⠁⠂⠄⡀⢀⠠⠐⠈ "),
-    );
-    scan_pb.set_message(format!("Scanning source directory: {}", dir_a));
-
-    // Collect all files in directory A
-    let mut file_paths = Vec::new();
-    let mut traversal_errors = 0;
-
-    for entry in WalkDir::new(dir_a) {
-        scan_pb.tick();
-        match entry {
-            Ok(dir_entry) => {
-                if dir_entry.file_type().is_file() {
-                    match Utf8PathBuf::from_path_buf(dir_entry.path().to_path_buf()) {
-                        Ok(utf8_path) => {
-                            file_paths.push(utf8_path);
-                            scan_pb.set_message(format!("Scanning... found {} files", file_paths.len()));
-                        }
-                        Err(path_buf) => {
-                            eprintln!("ERROR: Non-UTF8 path encountered: {:?}", path_buf);
-                            traversal_errors += 1;
-                        }
-                    }
-                }
-            }
-            Err(err) => {
-                eprintln!("ERROR: Directory traversal failed for source");
-                if let Some(path) = err.path() {
-                    eprintln!("  Path: {:?}", path);
-                }
-                eprintln!("  Error: {}", err);
-                traversal_errors += 1;
-            }
-        }
-    }
-
-    if traversal_errors > 0 {
-        eprintln!(
-            "WARNING: {} errors during source directory traversal",
-            traversal_errors
-        );
-    }
-
-    scan_pb.finish_with_message(format!("✓ Found {} files to copy", file_paths.len()));
-
-    // Phase 2: File processing with progress bar
-    let process_pb = multi.add(ProgressBar::new(file_paths.len() as u64));
-    process_pb.set_style(
-        ProgressStyle::with_template("📋 [{elapsed}] {bar:40.cyan/blue} {pos:>7}/{len:7} {msg}")
-            .unwrap()
-            .progress_chars("##-"),
-    );
-    process_pb.set_message("Processing files...");
-
-    let mut copied_count = 0;
-    let mut skipped_count = 0;
-    let mut error_count = 0;
-
-    for file_path in file_paths {
-        match process_single_file(&file_path, dir_a, dir_b, db, dry_run).await {
-            Ok(CopyResult::Copied) => {
-                copied_count += 1;
-                process_pb.set_message(format!("Copied: {} files, {} skipped, {} errors", 
-                    copied_count, skipped_count, error_count));
-                println!("COPIED: {}", file_path);
-            }
-            Ok(CopyResult::Skipped(reason)) => {
-                skipped_count += 1;
-                process_pb.set_message(format!("Copied: {} files, {} skipped, {} errors", 
-                    copied_count, skipped_count, error_count));
-                println!("SKIPPED: {} ({})", file_path, reason);
-            }
-            Err(e) => {
-                error_count += 1;
-                process_pb.set_message(format!("Copied: {} files, {} skipped, {} errors", 
-                    copied_count, skipped_count, error_count));
-                eprintln!("ERROR: Failed to process {}", file_path);
-                eprintln!("  Details: {:?}", e);
-            }
-        }
-        process_pb.inc(1);
-    }
-
-    process_pb.finish_with_message(format!("✓ Copy complete: {} copied, {} skipped, {} errors", 
-        copied_count, skipped_count, error_count));
-
-    // Print summary
-    println!("\nCOPY OPERATION SUMMARY:");
-    println!("  Files copied: {}", copied_count);
-    println!("  Files skipped: {}", skipped_count);
-    println!("  Errors: {}", error_count);
-
-    if dry_run {
-        println!("  (This was a dry run - no files were actually copied)");
-    }
-
-    Ok(())
-}
-
-#[derive(Debug)]
-enum CopyResult {
-    Copied,
-    Skipped(String),
-}
-
-async fn process_single_file(
-    file_path: &Utf8PathBuf,
-    source_base: &Utf8PathBuf,
-    target_base: &Utf8PathBuf,
-    db: &Database,
-    dry_run: bool,
-) -> Result<CopyResult, AppError> {
-    // Calculate relative path from source base
-    let relative_path = file_path
-        .strip_prefix(source_base)
-        .change_context(AppError::Copy)
-        .attach_printable_lazy(|| {
-            format!("Failed to calculate relative path for: {}", file_path)
-        })?;
-
-    // Calculate target path
-    let target_path = target_base.join(relative_path);
-
-    // Check if target file already exists
-    if target_path.exists() {
-        return Ok(CopyResult::Skipped(
-            "target file already exists".to_string(),
-        ));
-    }
-
-    // Get file metadata and hash
-    let metadata = fs::metadata(file_path)
-        .change_context(AppError::IO)
-        .attach_printable_lazy(|| format!("Failed to get metadata for: {}", file_path))?;
-
-    let last_modified = metadata
-        .modified()
-        .change_context(AppError::IO)
-        .attach_printable_lazy(|| format!("Failed to get modification time for: {}", file_path))?;
-
-    let file_hash = calculate_file_hash(file_path)?;
-
-    // Check if this hash already exists in the database (duplicate)
-    let existing_files = db
-        .get_files_with_hash(&file_hash)
-        .await
-        .change_context(AppError::DB)
-        .attach_printable_lazy(|| format!("file={file_path}"))?;
-
-    if !existing_files.is_empty() {
-        // Track this as a duplicate file
-        let original_path = &existing_files[0]; // Use first match as original
-        let file_size = metadata.len();
-        
-        if let Err(e) = db
-            .track_duplicate_file(file_path, original_path, &file_hash, file_size)
-            .await
-        {
-            eprintln!(
-                "WARNING: Failed to track duplicate file {}: {:?}",
-                file_path, e
-            );
-        }
-        
-        return Ok(CopyResult::Skipped(format!(
-            "duplicate content (matches {})",
-            original_path
-        )));
-    }
-
-    if dry_run {
-        return Ok(CopyResult::Copied);
-    }
-
-    // Create target directory if it doesn't exist
-    if let Some(parent) = target_path.parent() {
-        fs::create_dir_all(parent)
-            .change_context(AppError::Copy)
-            .attach_printable_lazy(|| format!("Failed to create target directory: {}", parent))?;
-    }
-
-    // Copy the file
-    fs::copy(file_path, &target_path)
-        .change_context(AppError::Copy)
-        .attach_printable_lazy(|| format!("Failed to copy {} to {}", file_path, target_path))?;
-
-    // Store hash in database for the new file
-    let file_size = metadata.len();
-    if let Err(e) = db
-        .upsert_file_hash(&target_path, &file_hash, file_size, last_modified)
-        .await
-    {
-        eprintln!(
-            "WARNING: Failed to store hash for copied file {}: {:?}",
-            target_path, e
-        );
-    }
-
-    // Track this as a copied file
-    if let Err(e) = db
-        .track_copied_file(file_path, &target_path, &file_hash, file_size)
-        .await
-    {
-        eprintln!(
-            "WARNING: Failed to track copied file {} -> {}: {:?}",
-            file_path, target_path, e
-        );
-    }
-
-    Ok(CopyResult::Copied)
-}
-
-async fn cleanup_tracked_files(
-    db: &Database,
-    multi: &MultiProgress,
-    dry_run: bool,
-) -> Result<(), AppError> {
-    println!("Starting cleanup of tracked files");
-    if dry_run {
-        println!("DRY RUN MODE: No files will actually be deleted");
-    }
-
-    // Phase 1: Get all tracked files
-    let scan_pb = multi.add(ProgressBar::new_spinner());
-    scan_pb.set_style(
-        ProgressStyle::with_template("🔍 {spinner:.green} {wide_msg}")
-            .unwrap()
-            .tick_chars("⠁⠂⠄⡀⢀⠠⠐⠈ "),
-    );
-    scan_pb.set_message("Loading tracked files from database...");
-
-    let copied_files = db
-        .get_copied_files()
-        .await
-        .change_context(AppError::Cleanup)
-        .attach_printable("Failed to retrieve copied files")?;
-
-    let duplicate_files = db
-        .get_duplicate_files()
-        .await
-        .change_context(AppError::Cleanup)
-        .attach_printable("Failed to retrieve duplicate files")?;
-
-    scan_pb.finish_with_message(format!(
-        "✓ Found {} copied files and {} duplicate files to process",
-        copied_files.len(),
-        duplicate_files.len()
-    ));
-
-    let total_files = copied_files.len() + duplicate_files.len();
-    if total_files == 0 {
-        println!("No tracked files found for cleanup");
-        return Ok(());
-    }
-
-    // Phase 2: Process copied files
-    let process_pb = multi.add(ProgressBar::new(total_files as u64));
-    process_pb.set_style(
-        ProgressStyle::with_template("🗑️  [{elapsed}] {bar:40.red/yellow} {pos:>7}/{len:7} {msg}")
-            .unwrap()
-            .progress_chars("##-"),
-    );
-    process_pb.set_message("Processing copied files...");
-
-    let mut deleted_count = 0;
-    let mut skipped_count = 0;
-    let mut error_count = 0;
-
-    println!("\n=== PROCESSING COPIED FILES ===");
-    for copied_file in copied_files {
-        match cleanup_copied_file(&copied_file, db, dry_run).await {
-            Ok(CleanupResult::Deleted) => {
-                deleted_count += 1;
-                println!("DELETED: {} (target: {})", copied_file.source_path, copied_file.target_path);
-            }
-            Ok(CleanupResult::Skipped(reason)) => {
-                skipped_count += 1;
-                println!("SKIPPED: {} ({})", copied_file.source_path, reason);
-            }
-            Err(e) => {
-                error_count += 1;
-                eprintln!("ERROR: Failed to cleanup {}: {:?}", copied_file.source_path, e);
-            }
-        }
-        process_pb.inc(1);
-        process_pb.set_message(format!("Cleaned: {} deleted, {} skipped, {} errors", 
-            deleted_count, skipped_count, error_count));
-    }
-
-    println!("\n=== PROCESSING DUPLICATE FILES ===");
-    for duplicate_file in duplicate_files {
-        match cleanup_duplicate_file(&duplicate_file, db, dry_run).await {
-            Ok(CleanupResult::Deleted) => {
-                deleted_count += 1;
-                println!("DELETED: {} (duplicate of: {})", duplicate_file.duplicate_path, duplicate_file.original_path);
-            }
-            Ok(CleanupResult::Skipped(reason)) => {
-                skipped_count += 1;
-                println!("SKIPPED: {} ({})", duplicate_file.duplicate_path, reason);
-            }
-            Err(e) => {
-                error_count += 1;
-                eprintln!("ERROR: Failed to cleanup {}: {:?}", duplicate_file.duplicate_path, e);
-            }
-        }
-        process_pb.inc(1);
-        process_pb.set_message(format!("Cleaned: {} deleted, {} skipped, {} errors", 
-            deleted_count, skipped_count, error_count));
-    }
-
-    process_pb.finish_with_message(format!("✓ Cleanup complete: {} deleted, {} skipped, {} errors", 
-        deleted_count, skipped_count, error_count));
-
-    // Print summary
-    println!("\nCLEANUP OPERATION SUMMARY:");
-    println!("  Files deleted: {}", deleted_count);
-    println!("  Files skipped: {}", skipped_count);
-    println!("  Errors: {}", error_count);
-
-    if dry_run {
-        println!("  (This was a dry run - no files were actually deleted)");
-    }
-
-    Ok(())
-}
-
-#[derive(Debug)]
-enum CleanupResult {
-    Deleted,
-    Skipped(String),
-}
-
-async fn cleanup_copied_file(
-    copied_file: &CopiedFileRecord,
-    db: &Database,
-    dry_run: bool,
-) -> Result<CleanupResult, AppError> {
-    // Check if target file still exists to verify the copy was successful
-    if !copied_file.target_path.exists() {
-        return Ok(CleanupResult::Skipped(
-            "target file no longer exists - copy may have failed".to_string(),
-        ));
-    }
-
-    // Check if source file still exists
-    if !copied_file.source_path.exists() {
-        // Remove tracking record since source is already gone
-        if !dry_run {
-            if let Err(e) = db
-                .remove_copied_file_record(&copied_file.source_path, &copied_file.target_path)
-                .await
-            {
-                eprintln!(
-                    "WARNING: Failed to remove tracking record for {}: {:?}",
-                    copied_file.source_path, e
-                );
-            }
-        }
-        return Ok(CleanupResult::Skipped(
-            "source file already deleted".to_string(),
-        ));
-    }
-
-    if dry_run {
-        return Ok(CleanupResult::Deleted);
-    }
-
-    // Delete the source file
-    fs::remove_file(&copied_file.source_path)
-        .change_context(AppError::Cleanup)
-        .attach_printable_lazy(|| {
-            format!("Failed to delete source file: {}", copied_file.source_path)
-        })?;
-
-    // Remove tracking record
-    if let Err(e) = db
-        .remove_copied_file_record(&copied_file.source_path, &copied_file.target_path)
-        .await
-    {
-        eprintln!(
-            "WARNING: Failed to remove tracking record for {}: {:?}",
-            copied_file.source_path, e
-        );
-    }
-
-    Ok(CleanupResult::Deleted)
-}
-
-async fn cleanup_duplicate_file(
-    duplicate_file: &DuplicateFileRecord,
-    db: &Database,
-    dry_run: bool,
-) -> Result<CleanupResult, AppError> {
-    // Check if duplicate file still exists
-    if !duplicate_file.duplicate_path.exists() {
-        // Remove tracking record since duplicate is already gone
-        if !dry_run {
-            if let Err(e) = db
-                .remove_duplicate_file_record(&duplicate_file.duplicate_path)
-                .await
-            {
-                eprintln!(
-                    "WARNING: Failed to remove tracking record for {}: {:?}",
-                    duplicate_file.duplicate_path, e
-                );
-            }
-        }
-        return Ok(CleanupResult::Skipped(
-            "duplicate file already deleted".to_string(),
-        ));
-    }
-
-    if dry_run {
-        return Ok(CleanupResult::Deleted);
-    }
-
-    // Delete the duplicate file
-    fs::remove_file(&duplicate_file.duplicate_path)
-        .change_context(AppError::Cleanup)
-        .attach_printable_lazy(|| {
-            format!("Failed to delete duplicate file: {}", duplicate_file.duplicate_path)
-        })?;
-
-    // Remove tracking record
-    if let Err(e) = db
-        .remove_duplicate_file_record(&duplicate_file.duplicate_path)
-        .await
-    {
-        eprintln!(
-            "WARNING: Failed to remove tracking record for {}: {:?}",
-            duplicate_file.duplicate_path, e
-        );
-    }
-
-    Ok(CleanupResult::Deleted)
-}
-
-async fn show_status(
-    db: &Database,
-    multi: &MultiProgress,
-    detailed: bool,
-) -> Result<(), AppError> {
-    println!("Retrieving status of files tracked for cleanup...");
-
-    // Get tracked files from database with spinner
-    let scan_pb = multi.add(ProgressBar::new_spinner());
-    scan_pb.set_style(
-        ProgressStyle::with_template("🔍 {spinner:.green} {wide_msg}")
-            .unwrap()
-            .tick_chars("⠁⠂⠄⡀⢀⠠⠐⠈ "),
-    );
-    scan_pb.set_message("Loading tracked files from database...");
-
-    let copied_files = db
-        .get_copied_files()
-        .await
-        .change_context(AppError::DB)
-        .attach_printable("Failed to retrieve copied files")?;
-
-    let duplicate_files = db
-        .get_duplicate_files()
-        .await
-        .change_context(AppError::DB)
-        .attach_printable("Failed to retrieve duplicate files")?;
-
-    scan_pb.finish_with_message("✓ Status retrieved");
-
-    // Calculate statistics
-    let copied_count = copied_files.len();
-    let duplicate_count = duplicate_files.len();
-    let total_files = copied_count + duplicate_count;
-
-    let copied_total_size: u64 = copied_files.iter().map(|f| f.file_size).sum();
-    let duplicate_total_size: u64 = duplicate_files.iter().map(|f| f.file_size).sum();
-    let total_size = copied_total_size + duplicate_total_size;
-
-    // Display summary
-    println!("\n=== CLEANUP STATUS SUMMARY ===");
-    println!("Files that would be deleted by cleanup command:");
-    println!("  Copied files (source files): {}", copied_count);
-    println!("  Duplicate files: {}", duplicate_count);
-    println!("  Total files: {}", total_files);
-    
-    if total_size > 0 {
-        println!("  Total size: {} bytes ({:.2} MB)", total_size, total_size as f64 / (1024.0 * 1024.0));
-    }
-
-    if total_files == 0 {
-        println!("\n✓ No files are currently tracked for cleanup");
-        return Ok(());
-    }
-
-    // Show detailed listing if requested
-    if detailed {
-        if copied_count > 0 {
-            println!("\n=== COPIED FILES (source files that would be deleted) ===");
-            for copied_file in &copied_files {
-                println!("  {} -> {}", copied_file.source_path, copied_file.target_path);
-            }
-        }
-
-        if duplicate_count > 0 {
-            println!("\n=== DUPLICATE FILES (that would be deleted) ===");
-            for duplicate_file in &duplicate_files {
-                println!("  {} (duplicate of: {})", duplicate_file.duplicate_path, duplicate_file.original_path);
-            }
-        }
-    } else if total_files > 0 {
-        println!("\nUse --detailed to see the full list of file paths that would be deleted.");
-    }
-
-    Ok(())
-}
-
 #[tokio::main]
 async fn main() -> Result<(), AppError> {
     let args = Args::parse();
 
     // Setup progress bars and logging
-    let multi = setup_progress_logging();
+    let multi = ProgressManager::setup();
 
     // Initialize database
     let db = Database::new(&args.db_path)
@@ -879,7 +80,8 @@ async fn main() -> Result<(), AppError> {
     match args.command {
         Commands::Hash { target_dir } => {
             println!("Running hash command for target directory: {}", target_dir);
-            scan_and_hash_directory(&target_dir, &db, &multi, args.batch_size).await?;
+            let scanner = FileScanner::new(args.batch_size);
+            scanner.scan_and_hash_directory(&target_dir, &db, &multi).await?;
             println!("Directory scanning and hashing completed");
         }
         Commands::Copy {
@@ -888,17 +90,20 @@ async fn main() -> Result<(), AppError> {
             dry_run,
         } => {
             println!("Running copy command");
-            validate_copy_command(&source, &target)?;
-            copy_files_without_duplicates(&source, &target, &db, &multi, dry_run).await?;
+            FileCopier::validate_copy_command(&source, &target)?;
+            let copier = FileCopier::new();
+            copier.copy_files_without_duplicates(&source, &target, &db, &multi, dry_run).await?;
             println!("Copy operation completed");
         }
         Commands::Cleanup { dry_run } => {
             println!("Running cleanup command");
-            cleanup_tracked_files(&db, &multi, dry_run).await?;
+            let cleaner = FileCleaner::new();
+            cleaner.cleanup_tracked_files(&db, &multi, dry_run).await?;
             println!("Cleanup operation completed");
         }
         Commands::Status { detailed } => {
-            show_status(&db, &multi, detailed).await?;
+            let cleaner = FileCleaner::new();
+            cleaner.show_status(&db, &multi, detailed).await?;
         }
     }
